api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
supabase>=2.4.0
python-dotenv>=1.0.1
websockets>=13.0
